% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accionescolestivas.R
\docType{data}
\name{accionescolestivas}
\alias{accionescolestivas}
\title{Acciones Colectivas}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 500 rows and 50 columns.
}
\usage{
data(accionescolestivas)
}
\description{
Estudio busca analizar las consecuencias psico-sociales y 
conductuales asociados a la participación directa, indirecta 
(vía identificación o apoyo) o no participación en movimientos
sociales y acciones colectivas y sus mecanismos de socialización
familiar (transmisión intergeneracional) en jóvenes y 
adultos pertenecientes a distintas generaciones en Chile.
}
\details{
Muestra: 500 estudiantes de distintas carreras y universidades de Santiago (250 mujeres y 250 hombres).
}
\examples{
data(accionescolestivas)
dplyr::glimpse((accionescolestivas)
}
\references{
Cornejo, M., Rocha, C., Castro, D., Varela, M., Manzi, J., 
González, R., Jiménez-Moya, G., Carvacho, H., Álvarez, B., Valdenegro, D., 
Cheyre, M., & Livingstone, A. G. (2021). 
The intergenerational transmission of participation in collective action: 
The role of conversation and political practices in the family.
The British Journal of Social, 60(1), 29-49.
https://doi.org/10.1111/bjso.12420
}
\keyword{datasets}
