% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amistad_intergrupal.r
\docType{data}
\name{amistad_intergrupal}
\alias{amistad_intergrupal}
\title{Amistad Integrupal}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{

}
\usage{
data(amistad_intergrupal)
}
\description{
Muestra de 300 estudiantes de educación Media de Santiago y Temuco.
Estudio busca explorar la influencia del contacto intergrupal en la promoción de mejores actitudes hacia las minorías étnicas, en contextos escolares.
Estudio observacional correlacional.
}
\examples{
data(amistad_intergrupal)
dplyr::glimpse((amistad_intergrupal)
}
\references{

}
\keyword{datasets}
