% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b5.R
\docType{data}
\name{b5}
\alias{b5}
\title{Inventario de personalidad B5}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
FONDECYT informacion por confirmar (20210419)
}
\usage{
data(b5)
}
\description{
Muestra de 151 personas, que respuestas a 60 items del big five
(afabilidad, apertura, extroversion, meticulosidad, neuroticismo).
}
\examples{
data(b5)
dplyr::glimpse(b5)
}
\references{
Benet-Martínez, V., & John, O. P. (1998). Los Cinco Grandes Across Cultures and Ethnic Groups: Multitrait Multimethod Analyses of the Big Five in Spanish and English. Journal of Personality and Social Psychology, 75(3), 729–750. https://doi.org/10.1037/0022-3514.75.3.729
}
\keyword{datasets}
