% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coes_2014.R
\docType{data}
\name{coes_2014}
\alias{coes_2014}
\title{Encuesta sobre Conflicto y Cohesión Social 2014}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
https://coes.cl/categoria/noticias-coes-encuesta-coes/
}
\usage{
data(coes_2014)
}
\description{
Estudio correlacional en base a encuesta aplicada a nivel nacional a adultos.
La muestra es una selección de personas residentes en la
Región Metropolitana de Santiago.
}
\details{
El objetivo de la encuesta ENACOES es aportar a la comprensión de las
creencias, actitudes y percepciones de los chilenos hacia
las distintas dimensiones de la convivencia y el conflicto.

La muestra total es de 2025 encuestados que residen
en Chile, de los cuáles se han seleccionado 225 (residentes en la Región
Metropolitana de Santiago) para efectos del trabajo en clases.
}
\examples{
data(coes_2014)
dplyr::glimpse((coes_2014)
}
\references{

}
\keyword{datasets}
