% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dem.R
\docType{data}
\name{data_dem}
\alias{data_dem}
\title{ICCS 2016 perfiles de concepciones de democracia}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
IEA, <https://www.iea.nl/index.php/data-tools/repository/iccs>
}
\usage{
data(data_dem)
}
\description{
Datos del Estudio Internacional sobre Educación Cívica y Ciudadana (ICCS).
Esta tabla de datos, incluye las observaciones del módulo internacional, y
del modulo latino americano.
Estos datos inlcuyen 500 casos de cada país.
}
\details{
Las variables incluidas son:

| variable  |type | labels                                                                                 |
| :---------|:--  | :--------------------------------------------------------------------------------------|
| id_i      |dbl  | id unico del caso                                                                      |
| ctry      |chr  | nombre del país                                                                        |
| sex       |dbl  | sexo del estudiante (0 = niño, 1 = niña)                                               |
| age       |dbl  | edad del estudiante en años (media = 14.25, sd = .99                                   |
| dem       |dbl  | perfil de concepciones de las democracia (1 = minimalista, 2 = complejo, 3 = limitado) |
| dem_group |chr  | perfil de concepciones de las democracia (en texto)                                    |
| aut       |dbl  | creencias antidemocráticas (AUTGOV, media =  50, sd = 10)                              |
| civ       |dbl  | conocimiento cívico (PV1CVI, media =  500, sd = 100)                                   |
| ses       |dbl  | nivel socioecónomico (media = 0, sd = 1, en cada país                                  |
| edu       |dbl  | educación de los padres (1 = educación universitaria, 0 = otro nivel educativo         |
}
\examples{
data(data_dem)
dplyr::glimpse(data_dem)
}
\references{
Schulz, W., Carstens, R., Losito, B., & Fraillon, J. (2018). ICCS 2016 Technical Report (W. Schulz, R. Carstens, B. Losito, & J. Fraillon (eds.)). International Association for the Evaluation of Educational Achievement (IEA).
}
\keyword{datasets}
