% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpi3.R
\docType{data}
\name{elpi3}
\alias{elpi3}
\title{ELPI III}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10698 rows and 153 columns.
}
\source{
http://observatorio.ministeriodesarrollosocial.gob.cl/elpi-tercera-ronda
}
\usage{
data(elpi)
}
\description{
Datos de la tercera encuesta longitudinal de primera infancia.
Se produce como el seguimiento de una muestra compleja y 
longitudinal, generada con información del registro civil de niñas y niños de Chile.
Incluye diferentes medidas de condiciones de nacimiento, 
estado de salud, nutrición, y otros aspectos del 
desarollo de los participantes, y su entorno familiar.
}
\details{
Para más detalles, consultar http://observatorio.ministeriodesarrollosocial.gob.cl/elpi-tercera-ronda
}
\examples{
data(elpi3)
dplyr::glimpse((elpi3)
}
\references{
http://observatorio.ministeriodesarrollosocial.gob.cl/elpi-tercera-ronda
}
\keyword{datasets}
