% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homonegatividad.R
\docType{data}
\name{homonegatividad}
\alias{homonegatividad}
\title{Homonegatividad}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 545 rows and 47 columns.
}
\source{
"..."
}
\usage{
data(homonegatividad)
}
\description{
Datos de actitudes hacia población LGBTQ+ por parte de jóvenes.
}
\examples{
data(homonegatividad)
dplyr::glimpse((homonegatividad)
}
\references{
"..."
}
\keyword{datasets}
