% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsb.R
\docType{data}
\name{hsb}
\alias{hsb}
\title{HSB data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7185 rows and 10 columns.
}
\source{
http://doi.wiley.com/10.1111/jasp.12322
}
\usage{
data(hsb)
}
\description{
Datos del artículo "Rehabilitating civilian victims of war through psychosocial intervention in Sierra Leone",
de Mughal, Carrasco, Brown & Ayers (2015). Este estudio evalua los niveles de estres post traumatico (PTSD),
de adultos en tratamiento en Sierra Leona. El grupo en tratamiento (n=50), fue reclutado por equipos de la
Universidad de Makeni; mientras que el grupo de control, eran adultos en lista de espera del
Holy Trinity Hospital.
}
\details{
Para crear los puntajes de variables, ver:
https://github.com/dacarras/psi2301_examples/blob/master/data/ptsd_add_mean_scores.rmd
}
\examples{
data(hsb)
dplyr::glimpse((hsb)
}
\references{
Mughal, U., Carrasco, D., Brown, R., & Ayers, S. (2015).
Rehabilitating civilian victims of war through psychosocial
intervention in Sierra Leone. Journal of Applied Social Psychology,
45(11), 593–601. https://doi.org/10.1111/jasp.12322
}
\keyword{datasets}
