% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccs_09_chl.R
\docType{data}
\name{iccs_09_chl}
\alias{iccs_09_chl}
\title{ICCS 2009 chl}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://www.iea.nl/index.php/data-tools/repository/iccs}
}
\usage{
data(iccs_09_chl)
}
\description{
Datos del Estudio Internacional sobre Educación Cívica y Ciudadana (ICCS).
Esta tabla de datos, incluye las observaciones del módulo latino american
de Chile de 2009.
}
\examples{
data(iccs_09_chl)
dplyr::glimpse((iccs_09_chl)
}
\references{

}
\keyword{datasets}
