% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccs_09_lat.R
\docType{data}
\name{iccs_09_lat}
\alias{iccs_09_lat}
\title{ICCS 2009 lat}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
IEA, <https://www.iea.nl/index.php/data-tools/repository/iccs>
}
\usage{
data(iccs_09_lat)
}
\description{
Datos del Estudio Internacional sobre Educación Cívica y Ciudadana (ICCS).
Esta tabla de datos, incluye las observaciones del módulo latino americano.
Estos datos inlcuyen muestras representativas de estudiantes de octavo grado,
de Chile, Colombia, México, República Dominicana, Guatemala, Mexico y Paraguay
}
\examples{
data(iccs_09_lat)
dplyr::glimpse(iccs_09_lat)
}
\references{
Schulz, W., Ainley, J., & Fraillon, J. (2011). ICCS 2009 Technical Report (W. Schulz, J. Ainley, & J. Fraillon (eds.)). International Association for the Evaluation of Educational Achievement (IEA). http://www.iea.nl/fileadmin/user_upload/Publications/Electronic_versions/ICCS_2009_Technical_Report.pdf
}
\keyword{datasets}
