% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iccs_16_lat}
\alias{iccs_16_lat}
\title{ICCS 2016}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(iccs_2016_lat)
}
\description{
Datos del Estudio Internacional de Educación Cívica ICCS 2016,
del modulo latinoamericano.
}
\examples{
data(iccs_16_lat)
dplyr::glimpse(iccs_16_lat)
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
