% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingresos_2016.R
\docType{data}
\name{ingresos_2016}
\alias{ingresos_2016}
\title{Ingresos 2016}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
INE, <https://www.ine.cl/estadisticas/sociales/ingresos-y-gastos/encuesta-de-presupuestos-familiares>
}
\usage{
data(ingresos_2016)
}
\description{
Muestra de 1120 casos aleatorios (20 caso por estrato), de los ingresos per capita,
provenientes de la base de datos "Encuesta de Presupuestos Familiares" (2016)
Incluye los datos de ingresos per capita por hogar en unidad de 1000,
sexo de los participantes, años de escolaridad alcanzada, y cantidad de personas
que comparten un hogar.
}
\examples{
data(ingresos_2016)
dplyr::glimpse((ingresos_2016)
}
\references{

}
\keyword{datasets}
