% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injuv_2012.R
\docType{data}
\name{injuv_2012}
\alias{injuv_2012}
\title{ENCUESTA INJUV 2012}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
https://www.injuv.gob.cl/7encuestanacionaldejuventud
}
\usage{
data(ingresos_2016)
}
\description{
Estudio correlacional en base a encuesta aplicada a nivel nacional a jóvenes entre 15 y 29 años
El objetivo de la encuesta del INJUV es tener una visión poblacional de diferentes temas de interés para los jóvenes en Chile
La muestra total es de más de 8000 encuestados, de los cuáles se han seleccionado 180 para efectos del trabajo en clases.
}
\examples{
data(ingresos_2016)
dplyr::glimpse((ingresos_2016)
}
\references{

}
\keyword{datasets}
