% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integracion_univ.R
\docType{data}
\name{integracion_universitaria}
\alias{integracion_universitaria}
\title{integracion_universitaria}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{

}
\usage{
data(ingresos_2016)
}
\description{
La base de datos se basa en una sub-muestra de 250 alumnos de la universida
que contestaron una encuesta enfocada en distintos ámbitos de su experiencia
de integración en la universidad. El cuestionario incluye preguntas
sobre las siguientes dimensiones:
}
\examples{
data(integracion_universitaria)
dplyr::glimpse((integracion_universitaria)
}
\references{
Nevill, A., & Rhodes, C. (2004). Academic and social integration
in higher education: a survey of satisfaction and dissatisfaction within a
first-year education studies cohort at a new university.
Journal of Further and Higher Education, 28(2), 179–193
}
\keyword{datasets}
