% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_2000.R
\docType{data}
\name{life_2000}
\alias{life_2000}
\title{life_2000}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
https://flowingdata.com/2011/10/13/life-expectancy-changes/
}
\usage{
data(life_2000)
}
\description{
Datos de expectativa de vida al nacer de 2000, y
producto interno bruto per carpita 'gross domestic produc)'.
Datos son obtenidos de "https://data.worldbank.org/" basados en
"https://flowingdata.com/2011/10/13/life-expectancy-changes/"
}
\examples{
data(life_2000)
dplyr::glimpse((life_2000)
}
\keyword{#'}
\keyword{datasets}
