% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptsd_data.R
\docType{data}
\name{ptsd_data}
\alias{ptsd_data}
\title{PTSD data}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
http://doi.wiley.com/10.1111/jasp.12322
}
\usage{
data(ptsd_data)
}
\description{
Datos del artículo "Rehabilitating civilian victims of war through psychosocial intervention in Sierra Leone",
de Mughal, Carrasco, Brown & Ayers (2015). Este estudio evalua los niveles de estres post traumatico (PTSD),
de adultos en tratamiento en Sierra Leona. El grupo en tratamiento (n=50), fue reclutado por equipos de la
Universidad de Makeni; mientras que el grupo de control, eran adultos en lista de espera del 
Holy Trinity Hospital.
}
\examples{
data(ptsd_data)
dplyr::glimpse((ptsd_data)
}
\references{
Mughal, U., Carrasco, D., Brown, R., & Ayers, S. (2015). 
Rehabilitating civilian victims of war through psychosocial 
intervention in Sierra Leone. Journal of Applied Social Psychology, 
45(11), 593–601. https://doi.org/10.1111/jasp.12322
}
\keyword{datasets}
