% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somking.R
\docType{data}
\name{smoking}
\alias{smoking}
\title{Smoking on birth outcomes}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
https://www.stata-press.com/data/mlmus3.html
}
\usage{
data(smoking)
}
\description{
Estudio panel con el peso de recien nacidos, y antecedentes de las madres,
incluyendo edad al nacer, consumo de cigarrillos por parte de la madre,
educación de la madres, entre otros indicadores.
}
\details{
Estos datos son empleados en Rabe-Hesketh & Skrondal (2012).
}
\examples{
data(smoking)
dplyr::glimpse((smoking)
}
\references{
Abrevaya, J. 2006. Estimating the eﬀect of smoking on birth outcomes using a matched panel data approach. Journal of Applied Econometrics 21: 489–519.
}
\keyword{datasets}
