% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terce.R
\docType{data}
\name{terce}
\alias{terce}
\title{terce}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
https://es.unesco.org/fieldoffice/santiago/llece/TERCE2013
}
\usage{
data(terce)
}
\description{
Este estudio evalúa el desempeño de estudiantes de tercer y sexto grado de
escuela primaria en las áreas de Matemática, Lectura y Escritura (Lenguaje),
y Ciencias Naturales en el caso de sexto grado.
}
\details{
Su objetivo principal es dar cuenta de la calidad de la educación en
Latinoamérica y guiar la toma de decisiones en políticas públicas educativas.
Para cumplir con este objetivo, el estudio no sólo consiste en la aplicación
de pruebas para medir logros de aprendizaje, sino también de cuestionarios
de contexto, para entender las circunstancias bajo las cuales el aprendizaje
ocurre.

La muestra efectiva del TERCE es de 3.065 escuelas en 15 países (Argentina,
Brasil, Chile, Colombia, Costa Rica, Ecuador, Guatemala, Honduras, México,
Nicaragua, Panamá, Paraguay, Perú, República Dominicana y Uruguay,
más el estado mexicano de Nuevo León).

Un total de 195.752 estudiantes participaron, de los cuales 100.752 fueron
de tercer grado y 95 mil de sexto grado.
sta base de datos es una submuestra de 400 participantes de Chile.
}
\examples{
data(terce)
dplyr::glimpse((terce)
}
\references{

}
\keyword{datasets}
