% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terremoto.r
\docType{data}
\name{terremoto}
\alias{terremoto}
\title{Terremoto}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
sin referencias
}
\usage{
data(terremoto)
}
\description{
Esta base contiene algunas variables que fueron medidas en una muestra de
adolescentes después del terremoto del año 2010,
que a continuación se describen:
}
\details{
mujer: 0= hombre, 1=mujer
edad: años
tept: diagnóstico de estrés postraumático, 0=ausencia, 1=presencia
autoef: escala de sentido de eficacia personal, mínimo=10, máximo= 35

Nota: datos empleados en taller 1, durante 2021.
}
\examples{
data(terremoto)
dplyr::glimpse(terremoto)
}
\references{
sin referencias
}
\keyword{datasets}
