% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccs_16_lat.R
\docType{data}
\name{iccs_16_lat}
\alias{iccs_16_lat}
\title{ICCS 2016 lat}
\format{
Objeto clase \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://www.iea.nl/index.php/data-tools/repository/iccs}
}
\usage{
data(iccs_16_lat)
}
\description{
Datos del Estudio Internacional sobre Educación Cívica y Ciudadana (ICCS).
Esta tabla de datos, incluye las observaciones del módulo latino americano.
Estos datos inlcuyen muestras representativas de estudiantes de octavo grado,
de Chile, Colombia, México, República Dominicana, y Perú.
}
\examples{
data(iccs_16_lat)
dplyr::glimpse(iccs_16_lat)
}
\references{
Schulz, W., Carstens, R., Losito, B., & Fraillon, J. (2018). ICCS 2016 Technical Report (W. Schulz, R. Carstens, B. Losito, & J. Fraillon (eds.)). International Association for the Evaluation of Educational Achievement (IEA).
}
\keyword{datasets}
